Fix: 변수 type 수정

See merge request s09-bigdata-recom-sub2/S09P22D206!34

+++ backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/CpuRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.pcgg.domain.recommend.entity.CpuEntity;
 
 @Repository
-public interface CpuRepository extends JpaRepository<CpuEntity, Integer> {
+public interface CpuRepository extends JpaRepository<CpuEntity, Long> {
 }
